优化代码

+++ src/main/java/cn/sakuramiku/lightblog/service/impl/TagServiceImpl.java
@@ -1,19 +1,22 @@
 package cn.sakuramiku.lightblog.service.impl;
 
-import cn.sakuramiku.lightblog.annotation.OnChange;
+import cn.sakuramiku.lightblog.annotation.OnCacheChange;
 import cn.sakuramiku.lightblog.common.annotation.LogConfig;
 import cn.sakuramiku.lightblog.common.annotation.WriteLog;
 import cn.sakuramiku.lightblog.common.util.IdGenerator;
 import cn.sakuramiku.lightblog.entity.Tag;
+import cn.sakuramiku.lightblog.exception.BusinessException;
 import cn.sakuramiku.lightblog.mapper.TagMapper;
 import cn.sakuramiku.lightblog.model.BatchInsertParam;
 import cn.sakuramiku.lightblog.service.TagService;
+import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
 import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.CacheEvict;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -34,6 +37,9 @@ public class TagServiceImpl implements TagService {
 
     @Resource
     private TagMapper tagMapper;
+    @Lazy
+    @Resource
+    private TagService tagService;
 
     @WriteLog(action = WriteLog.Action.INSERT)
     @CachePut(key = "#result.id",unless = "null  == #result")
@@ -47,7 +53,7 @@ public class TagServiceImpl implements TagService {
         tag.setName(name);
         Boolean insert = tagMapper.insert(tag);
         if (insert){
-            return this.getTag(id);
+            return getTag(id);
         }
         return null;
     }
@@ -59,7 +65,7 @@ public class TagServiceImpl implements TagService {
     public Tag updateTag(@NonNull Long id, @NonNull String name) {
         Boolean update = tagMapper.update(id, name);
         if (update){
-            return this.getTag(id);
+            return getTag(id);
         }
         return null;
     }
@@ -68,7 +74,11 @@ public class TagServiceImpl implements TagService {
     @CacheEvict(key = "#id")
     @Transactional(rollbackFor = Exception.class)
     @Override
-    public Boolean removeTag(@NonNull Long id) {
+    public Boolean removeTag(@NonNull Long id) throws BusinessException {
+        int articleCount = tagMapper.articleCount(id);
+        if (articleCount > 0){
+            throw new BusinessException("该标签还有文章在用哦");
+        }
         return tagMapper.delete(id);
     }
 
@@ -78,12 +88,18 @@ public class TagServiceImpl implements TagService {
         return tagMapper.get(id);
     }
 
-    @OnChange
+    @Override
+    public Tag getTagByName(String name) {
+        return tagMapper.getTagByName(name);
+    }
+
+    @OnCacheChange
     @Cacheable(unless = "null == #result || 0 == #result.total")
     @Override
     public PageInfo<Tag> search(Long articleId, String keyword, Integer page, Integer pageSize) {
         if (null != page && null != pageSize) {
-            PageHelper.startPage(page, pageSize, true);
+            Page<Object> objects = PageHelper.startPage(page, pageSize, true);
+            objects.setOrderBy("modified_time DESC");
         }
         List<Tag> tags;
         if (null != articleId) {
@@ -106,4 +122,10 @@ public class TagServiceImpl implements TagService {
     public Boolean batchInsert(List<BatchInsertParam> params) {
         return tagMapper.batchInsert(params);
     }
+
+    @Override
+    public Boolean deleteForArticle(Long articleId) {
+        return tagMapper.deleteForArticle(articleId);
+    }
+
 }
